[#1087] Renamed NotFoundPage.
